+++ backend/src/main/kotlin/com/infraView/alarm/domain/Alarm.kt
@@ -3,9 +3,9 @@ package com.infraView.alarm.domain
 import java.time.OffsetDateTime
 
 data class Alarm(
-    val id: Int? = null,
+    val id: Long? = null,
     val alarmType: AlarmType,
     val triggeredAt: OffsetDateTime,
     var resolvedAt: OffsetDateTime? = null,
-    val incidentId: Int
+    val incidentId: Long
 )
