+++ backend/src/main/kotlin/com/infraView/incident/domain/IncidentService.kt
@@ -13,7 +13,7 @@ class IncidentService(
         return incidentPort.getAll()
     }
 
-    override fun getById(id: Int): Incident? {
+    override fun getById(id: Long): Incident? {
         return incidentPort.getById(id)
     }
 
@@ -21,14 +21,19 @@ class IncidentService(
         return incidentPort.save(incident)
     }
 
-    override fun endIncident(incidentId: Int): Incident? {
+    override fun endIncident(incidentId: Long): Incident? {
+        // Wyciągamy czysty obiekt domenowy przez Port (zamiast starego repozytorium)
         val incident = incidentPort.getById(incidentId) ?: throw RuntimeException("Incident not found")
+        
+        // Zmieniamy stan naszej czystej domeny (bez DTO i Encji!)
         incident.endedAt = OffsetDateTime.now()
         incident.status = StatusType.RESOLVED
+
+        // Zapisujemy przez port wyjściowy i zwracamy zaktualizowaną domenę
         return incidentPort.save(incident)
     }
 
-    override fun delete(incidentId: Int) {
+    override fun delete(incidentId: Long) {
         incidentPort.deleteById(incidentId)
     }
 }
