+++ backend/src/main/kotlin/com/infraView/telemetry/database/SpringDataTelemetryRepository.kt
@@ -3,6 +3,6 @@ package com.infraView.telemetry.database
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface SpringDataTelemetryRepository : JpaRepository<TelemetryJpaEntity, Long> {
-    fun findAllByIncidentId(incidentId: Int): List<TelemetryJpaEntity>
-    fun deleteAllByIncidentId(incidentId: Int)
+    fun findAllByIncidentId(incidentId: Long): List<TelemetryJpaEntity>
+    fun deleteAllByIncidentId(incidentId: Long)
 }
